% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp21century}
\alias{gdp21century}
\title{GDP Data of the 21st Century}
\format{A data frame with 4752 rows and 11 variables:
\describe{
  \item{country}{country or region name}
  \item{NY.GDP.PCAP.PP.KD}{GDP per capita}
  \item{year}{year}
  ...
}}
\source{
\url{https://datacatalog.worldbank.org/}
}
\usage{
gdp21century
}
\description{
A dataset containing GDP per capita information for all countries and regions from
year 2000 to 2017
}
\keyword{datasets}
