% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_round.R
\name{pos_round}
\alias{pos_round}
\title{Helps users to keep data clean}
\usage{
pos_round(x, r = T)
}
\arguments{
\item{x}{A numeric vector.}

\item{r}{A logical variable to control whether to round the number or not.}
}
\value{
Positive numbers in \code{x}, rounded if \code{y} is TRUE.
}
\description{
Subset and round positive numbers in a numeric vector.
}
\examples{
pos_round(rnorm(10), TRUE)
pos_round(rnorm(5), FALSE)
}
